requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.13.2
anyio==4.12.0
apache-tvm-ffi==0.1.7
astor==0.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
beir==2.1.0
bitsandbytes==0.46.0
blake3==1.0.8
blis==1.3.0
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
cbor2==5.8.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.0
click==8.3.1
cloudpathlib==0.21.0
cloudpickle==3.1.2
comm==0.2.1
compressed-tensors==0.12.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.8.0
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
cycler==0.12.1
cymem==2.0.11
datasets==4.4.2
debugpy==1.8.1
decorator==5.1.1
depyf==0.20.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.21.2
einops==0.8.0
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
fastrlock==0.8.3
ffmpy==0.4.0
filelock==3.20.0
flash-attn==2.7.0.post2
flashinfer-python==0.5.2
fonttools==4.55.0
frozenlist==1.5.0
fschat==0.2.36
fsspec==2023.10.0
fuzzywuzzy==0.18.0
gguf==0.17.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
gradio==5.6.0
gradio_client==1.4.3
greenlet==3.3.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.0
idna==3.10
ijson==3.4.0.post0
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflect==7.3.1
interegular==0.3.3
ipykernel==6.29.3
ipython==8.22.1
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==3.12.1
jedi==0.19.1
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kernels==0.12.1
keyring==25.5.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
latex2mathml==3.77.0
latex2sympy2_extended==1.10.2
Levenshtein==0.27.3
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
math-verify==0.8.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mcp==1.25.0
mdurl==0.1.2
mistral_common==1.8.8
model-hosting-container-standards==0.1.12
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.18
ninja==1.13.0
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.4
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
orjson==3.10.11
outlines_core==0.2.11
packaging==25.0
pandas==2.2.3
parso==0.8.3
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pexpect==4.9.0
pillow==11.3.0
pkginfo==1.12.0
platformdirs==4.2.0
playwright==1.56.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt-toolkit==3.0.43
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pydyf==0.11.0
pyee==13.0.0
Pygments==2.17.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pynvml==11.5.3
pyparsing==3.2.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.14.3
ray==2.53.0
readme_renderer==44.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rich-toolkit==0.17.1
rignore==0.7.6
rouge==1.0.1
rpds-py==0.30.0
rsa==4.9.1
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.48.0
setproctitle==1.3.7
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==3.1.2
stack-data==0.6.3
starlette==0.50.0
supervisor==4.3.0
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.5.0
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers==4.56.2
triton==3.5.0
trl==0.26.2
twine==6.0.1
typeguard==4.3.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.22.1
vllm==0.11.2
wandb==0.23.1
wasabi==1.1.3
watchfiles==1.1.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.1
-e git+https://github.com/facebookresearch/wasp.git@ffee6f41fde76acd14bd792db442479c506260c2#egg=webarena&subdirectory=visualwebarena
webencodings==0.5.1
websockets==12.0
wrapt==1.17.2
xformers==0.0.33.post1
xgrammar==0.1.25
xxhash==3.5.0
yarl==1.17.2
zipp==3.21.0
zopfli==0.2.3.post1
